+++ checks.go
@@ -18,16 +18,16 @@ var AvailableChecks = make(map[string]CheckEntry)
 
 const (
 	// CheckExitOK is used for normal exits.
-	CheckExitOK = 0
+	CheckExitOK = int64(0)
 
 	// CheckExitWarning is used for warnings.
-	CheckExitWarning = 1
+	CheckExitWarning = int64(1)
 
 	// CheckExitCritical is used for critical errors.
-	CheckExitCritical = 2
+	CheckExitCritical = int64(2)
 
 	// CheckExitUnknown is used for when the check runs into a problem itself.
-	CheckExitUnknown = 3
+	CheckExitUnknown = int64(3)
 )
 
 // CheckResult is the result of a single check run.
